+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Let the build fail when bibtex finds no bib file.
+
 ## [0.4.0] - 2023-03-24
 
 ### Changed