+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-03-25
+
 ### Fixed
 
 - Add missing depfile lines to `latexdep` and `bibtex` build rules.